🔨 Apply Coding Conventions to CreateNewSandboxModal
Applied our coding conventions, reorganized, and converted most of the CreateNewSandbox files to TypeScript.

Was unabled to refactor CreateNewSandbox/index.js as I wasn't quite sure about how it works. @CompuIves you may want to take a shot at this or otherwise I'll just have to come back to it later.

Tested locally, all the functionality appears to be the same as before.